requirements.txt
streamlit 
onnxruntime
pillow
numpy 
pandas